requirements.txt
google-generativeai>=0.2.0
streamlit>=1.30.0
langchain>=0.1.0
langgraph>=0.0.10
python-dotenv>=1.0.0
duckduckgo-search>=3.9.0
